fastapi-backend/docker-scraper/requirements.txt
beautifulsoup4==4.12.2
PyMySQL==1.1.0
selenium==4.19.0
trio==0.25.0
trio-websocket==0.11.1
webdriver-manager==4.0.1
websockets==10.4
wsproto==1.2.0
python-dotenv==1.0.1
pytest
pytest-mock